+++ fim_pipeline.sh
@@ -95,7 +95,10 @@ process_wb_file=$projectDir/fim_process_huc.sh
 pipeline_start_time=`date +%s`
 
 # PROCESS THE UNITS (And branches)
-# Why an if and else? watch the number of colons
+# Why an if and else? .. watch the number of colons.
+echo
+echo "---- Unit (HUC) processing is started"
+
 if [ -f "$hucList" ]; then
     if [ "$jobHucLimit" = "1" ]; then
         parallel --verbose --lb -j $jobHucLimit --colsep ',' --joblog $logFile -- $process_wb_file $runName :::: $hucList